% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starwars.R
\name{starwars}
\alias{starwars}
\title{Star Wars}
\usage{
starwars(episode = "sw1", fps = 20)
}
\arguments{
\item{episode}{Episode to show. Currently, "sw1" and "sw2" are available.}

\item{fps}{Approximate frame per second.}
}
\description{
Flushing Star Wars episodes in R console
}
\examples{
#starwars("sw1", 20)
}
